chore: SDK packaging prep — Swift single-repo, descriptions, Rust publish unblock
Swift (one repo, root manifest):
- Add root Package.swift (targets path into swift/) so SPM consumes the
  monorepo by URL; delete nested swift/Package.swift. Repoint both harness
  clients and PayKitDemo.xcodeproj at the root package; update swift.yml
  (run at root, .build path) and the docs-swift recipe. README install
  snippet -> GitHub URL.

Repo cleanup:
- Remove redundant root package.json (playground scripts live in
  playground/; CI uses the typescript/ workspace).
- Rename justfile -> Justfile.

Descriptions -> "Building blocks for Agentic payments (x402, MPP, AP2)"
for the main package per language: Python, TS @solana/pay-kit, Rust kit,
PHP, Ruby (summary), Lua (summary), Swift (README tagline).

Rust publish unblock:
- solana-keychain: git rev -> crates.io "1.4" (1.4.0 supersedes the pinned
  rev) in core/mpp/x402.
- Add version requirements to inter-crate path deps so the tree is
  publishable.
