+++ .github/workflows/swift.yml
@@ -13,14 +13,15 @@ jobs:
     steps:
       - uses: actions/checkout@v5
       - name: Run tests with coverage
-        working-directory: swift
+        # The Swift package manifest lives at the repo root (so SPM can consume
+        # this monorepo by URL); sources are under swift/.
         run: swift test --enable-code-coverage
       - name: Upload Swift coverage
         if: always()
         uses: actions/upload-artifact@v7
         with:
           name: swift-coverage
-          path: swift/.build/debug/codecov/
+          path: .build/debug/codecov/
           if-no-files-found: ignore
 
   harness-swift:
