+++ Package.swift
@@ -1,0 +1,44 @@
+// swift-tools-version: 6.0
+
+import PackageDescription
+
+// Root manifest so Swift Package Manager can consume this monorepo by its
+// GitHub URL directly (SPM requires Package.swift at the repo root, and resolves
+// versions from bare-semver git tags). The Swift SDK sources live under swift/;
+// targets below point at them via `path:`.
+//
+//     .package(url: "https://github.com/solana-foundation/pay-kit.git", from: "0.6.1")
+//
+// API docs are generated by `sourcedocs` into docs/api/swift/ (`just docs-swift`).
+let package = Package(
+    name: "SolanaPayKit",
+    platforms: [
+        .iOS(.v16),
+        .macOS(.v13),
+    ],
+    products: [
+        .library(
+            name: "SolanaPayKit",
+            targets: ["SolanaPayKit"]
+        ),
+    ],
+    targets: [
+        .target(
+            name: "SolanaPayKit",
+            path: "swift/Sources/SolanaPayKit"
+        ),
+        // Cross-SDK conformance-vector runner CLI. Reads one vector as JSON on
+        // stdin and emits one RunnerResult line on stdout, per the contract in
+        // harness/src/conformance/ts-runner.ts.
+        .executableTarget(
+            name: "mpp-conformance",
+            dependencies: ["SolanaPayKit"],
+            path: "swift/Sources/mpp-conformance"
+        ),
+        .testTarget(
+            name: "SolanaPayKitTests",
+            dependencies: ["SolanaPayKit"],
+            path: "swift/Tests/SolanaPayKitTests"
+        ),
+    ]
+)
