+++ rust/crates/core/Cargo.toml
@@ -29,7 +29,7 @@ otel = [
 
 [dependencies]
 # Signing — solana-keychain with sdk-v3 (matches mpp/x402 pins)
-solana-keychain = { git = "https://github.com/solana-foundation/solana-keychain", rev = "d788028edbe02a94ef5eee7585d0230ad771296e", default-features = false, features = ["memory", "sdk-v3"] }
+solana-keychain = { version = "1.4", default-features = false, features = ["memory", "sdk-v3"] }
 
 # Force five8_core's `std` feature ON: it gates `impl Error for DecodeError`,
 # which solana-keypair/solana-signature rely on. Without it, building this crate
@@ -53,7 +53,7 @@ tracing = { version = "0.1", optional = true }
 async-trait = { version = "0.1", optional = true }
 
 # Payment-channels program client (generated)
-payment_channels_client = { package = "payment-channels-client", path = "../programs/payment-channels" }
+payment_channels_client = { package = "payment-channels-client", version = "0.1.0", path = "../programs/payment-channels" }
 
 # Randomness (channel salt)
 getrandom = "0.3"
