--- swift/Package.swift
@@ -1,37 +1,0 @@
-// swift-tools-version: 6.0
-
-import PackageDescription
-
-let package = Package(
-    name: "SolanaPayKit",
-    platforms: [
-        .iOS(.v16),
-        .macOS(.v13),
-    ],
-    products: [
-        .library(
-            name: "SolanaPayKit",
-            targets: ["SolanaPayKit"]
-        ),
-    ],
-    targets: [
-        .target(name: "SolanaPayKit"),
-        // Cross-SDK conformance-vector runner CLI. Reads one vector as JSON
-        // on stdin and emits one RunnerResult line on stdout, honoring the
-        // contract in harness/src/conformance/ts-runner.ts. Swift is a
-        // client-only SDK: it drives charge + x402-exact build paths and the
-        // canonical-bytes encoders, and emits unsupported-mode for the
-        // server verify vectors so the driver skips them.
-        .executableTarget(
-            name: "mpp-conformance",
-            dependencies: ["SolanaPayKit"]
-        ),
-        .testTarget(
-            name: "SolanaPayKitTests",
-            dependencies: ["SolanaPayKit"]
-        ),
-    ]
-)
-
-// API docs are generated by `sourcedocs` into ../docs/api/swift/ as markdown.
-// Run `just docs-swift` from the repo root.
